[#25] refactor: Controller 전역에서 발생하는 예외를 처리하도록 HandlerAdvice 클래스 추가
- @RestControllerAdvice 어노테이션을 사용하여 해결
